usb: fix StreamOn for non integer fps
BUG=chromium:852302,chromium:881300
TEST=Run capture_unittests with vivid on VM, which uses 1280x720 in 12.5
     fps.

Reviewed-on: https://chromium-review.googlesource.com/1286024
Commit-Ready: Keiichi Watanabe <keiichiw@chromium.org>
Tested-by: Keiichi Watanabe <keiichiw@chromium.org>

+++ camera/hal/usb/v4l2_camera_device.cc
@@ -170,8 +170,9 @@ int V4L2CameraDevice::StreamOn(uint32_t width,
                << streamparm.parm.capture.timeperframe.numerator;
     }
   }
-  float fps = streamparm.parm.capture.timeperframe.denominator /
-              streamparm.parm.capture.timeperframe.numerator;
+  float fps =
+      static_cast<float>(streamparm.parm.capture.timeperframe.denominator) /
+      streamparm.parm.capture.timeperframe.numerator;
   if (std::fabs(fps - frame_rate) > std::numeric_limits<float>::epsilon()) {
     LOGF(ERROR) << "Unsupported frame rate " << frame_rate;
     return -EINVAL;
